Update ec2_private_instance.md
The VPC, public subnet, private subnet, and internet gateway resources are created as before.
Separate key pairs are created for the bastion host and the private EC2 instance using the aws_key_pair resource. Make sure to replace the paths to the public key files (~/.ssh/bastion_key.pub and ~/.ssh/private_instance_key.pub) with the actual paths to your public key files.
The security groups for the bastion host and the private EC2 instance are created to allow SSH access.
The bastion host (aws_instance.bastion) is created in the public subnet, and the private EC2 instance (aws_instance.private_instance) is created in the private subnet.
The user data for the private EC2 instance includes the installation of the CloudWatch agent and configuration to send logs to a CloudWatch Logs group named "MyLogGroup".